fix(frontend): dedupe client-side search filters (MAS-565)
